회원정보  1차 수정 데이터 수정후 저장 기능 완료

+++ src/main/java/com/capstone/tutorlink/domain/member/command/domain/repository/MemberRepository.java
@@ -10,5 +10,5 @@ import java.util.Optional;
 public interface MemberRepository extends JpaRepository<Member, Integer> {
 
     Optional<Member> findByMemberIdAndMemberCurrentStatus(String memberId, String memberCurrentStatus);
-    Member findByMemberNo(Long memberNo);
+    Member findByMemberNo(int memberNo);
 }
